Create a generic FacetPaginator class, Solr::FacetPaginator sub-classes it.
